Refactor identifier utils in order to be accessible from the debugger
