+++ bvm/ballerina-runtime/src/main/java/org/ballerinalang/jvm/types/BType.java
@@ -17,7 +17,7 @@
 */
 package org.ballerinalang.jvm.types;
 
-import org.ballerinalang.jvm.IdentifierEncoder;
+import org.ballerinalang.jvm.IdentifierUtils;
 import org.ballerinalang.jvm.TypeChecker;
 import org.ballerinalang.jvm.util.exceptions.BallerinaException;
 
@@ -40,7 +40,7 @@ public abstract class BType {
     private int hashCode;
 
     protected BType(String typeName, BPackage pkg, Class<? extends Object> valueClass) {
-        this.typeName = IdentifierEncoder.decodeIdentifier(typeName);
+        this.typeName = IdentifierUtils.decodeIdentifier(typeName);
         this.pkg = pkg;
         this.valueClass = valueClass;
         if (pkg != null && typeName != null) {
